chore: Se modifica la carga de las tablas en todos los modulos
* Se hace uso de javascript para obtener por medio de un endpoint los datos de la tabla en lugar de cargar directamente en el html
* Se crean los endpoints datatables  para cada modulo que requiera listar por medio de una lista

+++ public/views/Categories/categories.js
@@ -1,4 +1,8 @@
-/* OBTENER EL USUARIO A EDITAR */
+$(document).ready(function () {
+	getDataForTable();
+});
+
+/* OBTENER LA CATEGORIA A EDITAR */
 $(document).on("click", ".btn-CategoryEdit", function (e) {
 	var categoryId = $(this).attr("categoryId");
 
@@ -15,25 +19,40 @@ $(document).on("click", ".btn-CategoryEdit", function (e) {
 		.catch((error) => alert.errorAlert(error));
 });
 
-/* ELIMINAR UN USUARIO */
+/* ELIMINAR UN CATEGORIA */
 $(document).on("click", ".btn-DeleteCategory", function (e) {
-	var categoryId = $(this).attr("categoryId");
+	Swal.fire({
+		title: "Estas seguro de eliminar este registro?",
+		icon: "warning",
+		showCancelButton: true,
+		confirmButtonColor: "#3085d6",
+		cancelButtonColor: "#d33",
+		confirmButtonText: "!Si, Estoy seguro!",
+		cancelButtonText: "Cancelar",
+	}).then((result) => {
+		if (!result.value) {
+			return;
+		}
 
-	fetch("./category/delete?id=" + categoryId, {
-		method: "DELETE",
-	})
-		.then((res) => res.json())
-		.then((response) => {
-			if (response.error) {
-				alert.errorAlert(response.message);
-			} else {
-				alert.successAlert(response.message);
-			}
+		var categoryId = $(this).attr("categoryId");
+
+		fetch("./category/delete?id=" + categoryId, {
+			method: "DELETE",
 		})
-		.catch((error) => alert.errorAlert(error));
+			.then((res) => res.json())
+			.then((response) => {
+				if (response.error) {
+					alert.errorAlert(response.message);
+				} else {
+					alert.successAlert(response.message);
+					$("#categoryTable").DataTable().ajax.reload();
+				}
+			})
+			.catch((error) => alert.errorAlert(error));
+	});
 });
 
-/* EDITAR UN USUARIO */
+/* EDITAR UN CATEGORIA */
 $(document).on("click", "#btnEditCategory", function (e) {
 	formData = document.getElementById("FormEditCategory").getData();
 
@@ -52,12 +71,13 @@ $(document).on("click", "#btnEditCategory", function (e) {
 				alert.errorAlert(response.message);
 			} else {
 				alert.successAlert(response.message);
+				$("#categoryTable").DataTable().ajax.reload();
 			}
 		})
 		.catch((error) => alert.errorAlert(error));
 });
 
-/* CREAR UN USUARIO */
+/* CREAR UN CATEGORIA */
 $(document).on("click", "#btnAddCategory", function (e) {
 	var data = {
 		...document.getElementById("FormAddCategory").getData(),
@@ -73,6 +93,7 @@ $(document).on("click", "#btnAddCategory", function (e) {
 				alert.errorAlert(response.message);
 			} else {
 				alert.successAlert(response.message);
+				$("#categoryTable").DataTable().ajax.reload();
 				document.getElementById("FormAddCategory").reset();
 			}
 		})
@@ -109,3 +130,74 @@ $(document).on("click", ".btn-Active", function (e) {
 				.catch((error) => alert.errorAlert(error));
 		});
 });
+
+function getDataForTable() {
+	fetch("./login/rol")
+		.then((response) => response.json())
+		.then(({ rolCode }) => {
+			table = $("#categoryTable");
+			
+			table.DataTable().destroy();
+
+			table.DataTable({
+				order: [[1, "asc"]],
+				processing: true,
+				language: {
+					url: "//cdn.datatables.net/plug-ins/1.10.15/i18n/Spanish.json",
+				},
+				ajax: {
+					type: "GET",
+					url: `./category/datatable`,
+				},
+
+				columns: [
+					{
+						data: "id",
+						render: function (data) {
+							return data;
+						},
+						visible: false,
+					},
+					{
+						data: "description",
+						render: function (data) {
+							return data;
+						},
+					},
+					{
+						data: "active",
+						render: function (data, type, row) {
+							return data == 1
+								? `<td class='cell'><button class='btn btn-success btn-xs btn-Active' categoryId='${row.id}'>Activo</button></td>`
+								: `<td class='cell'><button class='btn btn-danger btn-xs btn-Active' categoryId='${row.id}'>Inactivo</button></td>`;
+						},
+						className: "cell",
+					},
+					{
+						data: "",
+						render: function (data, type, row) {
+							return `
+							<td class='cell'>
+									<div class='btn-group'>
+										${
+											["Admin", "InventoryManager"].includes(rolCode)
+												? `
+												<button class='btn btn-warning text-white m-0 btn-CategoryEdit' categoryId='${row.id}' data-toggle='modal' data-target='#EditCategoryModal'>
+													<i class='fas fa-pencil-alt'></i>
+												</button>
+												<button class='btn btn-danger text-white m-0 btn-DeleteCategory' categoryId='${row.id}'>
+													<i class='fas fa-times'></i>
+												</button>`
+												: "No disponible"
+										}
+										
+									</div>
+								</td>
+							</tr>`;
+						},
+						className: "cell",
+					},
+				],
+			});
+		});
+}
